+++ changelog.md
@@ -27,6 +27,7 @@ For example:
 * Fixed a bug that make the grid view crash when searching text and a formula field is referencing a singe-select field. [#1110](https://gitlab.com/bramw/baserow/-/issues/1110)
 * Fixed horizontal scroll on Mac OSX.
 * Fixed bug where the row coloring didn't work in combination with group level premium.
+* Fixed a bug when inserting hundreds of rows all with the same before parameter that causes subsequent inserts with the same before to slow down. [#1083](https://gitlab.com/bramw/baserow/-/issues/1083)
 
 ### Refactors
 * Fix view and fields getting out of date on realtime updates. [#1112](https://gitlab.com/bramw/baserow/-/issues/1112)